prepare for sklearn release 0.24

+++ autosklearn/metalearning/metalearning/meta_base.py
@@ -1,3 +1,4 @@
+import numpy as np
 import pandas as pd
 
 from ...util.logging_ import get_logger
@@ -57,7 +58,7 @@ class MetaBase(object):
         metafeatures.name = name
         if isinstance(metafeatures, DatasetMetafeatures):
             data_ = {mf.name: mf.value for mf in metafeatures.metafeature_values.values()}
-            metafeatures = pd.Series(name=name, data=data_)
+            metafeatures = pd.Series(name=name, data=data_, dtype=np.float64)
         if name.lower() in self.metafeatures.index:
             self.logger.warning(
                 'Dataset %s already in meta-data. Removing occurence.', name
